pyproject.toml
[tool.poetry]
name = "gemini-stock-analysis"
version = "0.1.0"
description = "Stock analysis project using Gemini API, MCP, and vector databases with Google Sheets input"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "gemini_stock_analysis"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
google-generativeai = "^0.8.0"
google-api-python-client = "^2.150.0"
google-auth = "^2.43.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
chromadb = "^1.3.5"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.0"
numpy = "^1.26.0"
fastapi = "^0.123.0"
fastapi-mcp = "^0.4.0"
uvicorn = "^0.38.0"
httpx = "^0.28.1"
python-multipart = "^0.0.20"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.0"
mypy = "^1.11.0"

[tool.poetry.scripts]
gemini-stock-analysis = "gemini_stock_analysis.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

